+++ rust_mbu/rust_engine/engine/src/engine/nn/readcsv.rs
@@ -4,7 +4,7 @@ extern crate csv;
 use ndarray::{Array1, Array2};
 use std::fs::File;
 
-pub fn read_array2_from_csv(file_path: Option<&str>) -> Array2<f64> {
+pub fn read_array2_from_csv(file_path: Option<&str>) -> Array2<f32> {
     let file = read_csv(file_path.unwrap());
 
     let mut rdr = csv::ReaderBuilder::new().has_headers(false).from_reader(file);
@@ -17,12 +17,12 @@ pub fn read_array2_from_csv(file_path: Option<&str>) -> Array2<f64> {
             Err(_error) => panic!("Wrong record"),
         };
 
-        let row: Vec<f64> = record.iter().map(|x| x.parse::<f64>().unwrap()).collect();
+        let row: Vec<f32> = record.iter().map(|x| x.parse::<f32>().unwrap()).collect();
         records.push(row);
     }
     let rows = records.len();
     let cols = records[0].len();
-    let flat_data: Vec<f64> = records.into_iter().flatten().collect();
+    let flat_data: Vec<f32> = records.into_iter().flatten().collect();
     let array = Array2::from_shape_vec((rows, cols), flat_data);
 
     match array {
@@ -32,7 +32,7 @@ pub fn read_array2_from_csv(file_path: Option<&str>) -> Array2<f64> {
 }
 
 
-pub fn read_array1_from_csv(file_path: Option<&str>) -> Array1<f64> {
+pub fn read_array1_from_csv(file_path: Option<&str>) -> Array1<f32> {
     let file = read_csv(file_path.unwrap());
 
     let mut rdr = csv::ReaderBuilder::new().has_headers(false).from_reader(file);
@@ -45,7 +45,7 @@ pub fn read_array1_from_csv(file_path: Option<&str>) -> Array1<f64> {
             Err(_error) => panic!("Wrong record"),
         };
         for field in record.iter() {
-            records.push(field.parse::<f64>().unwrap());
+            records.push(field.parse::<f32>().unwrap());
         }
     }
     return Array1::from(records);
